==> R/1_nimbleModel_multiOcc_v11_farm.Rmd <==
---
  title: "bayesBirdModel_multiSPecies_v11 - FARM"
author: "Liz Law"
date: "`r Sys.Date()`"
output: html_document
---
  workingconservation@gmail.com

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Multi-species occupancy model using nimbleEcology
* marginalized occupancy likelihood (docc)
* uniform priors
* variables are normalised prior to analysis
* community occupancy hyperparameter sd.betalpsi not parameter specific
* community detection hyperparameter sd.betalp not parameter specific

# Setup
```{r library setup}
library(nimble)
library(nimbleEcology)
library(tidyverse)

mySeed <- 101 
set.seed(mySeed)

wd <- "/Users/elaw/Desktop/LeuphanaProject/BirdModelling/ETH_birds"
#wd <- "C:/Users/chloe.nater/OneDrive - NINA/Documents/Projects/Other/BirdOccupancy_Liz"

results_folder <- paste0(wd, "/Results/")  
version_folder <- "v11/"
data_version <- "v7"
data_input <- paste0(wd, "/Data/nimbleModel_multiOcc_", data_version,"_allData.RDS")

trialName <- "NoPolyFVars"  # hyperpriors priors inits library params manually altered. 
```

# build model
```{r model code}
occupancy_code <- nimbleCode({
  
  # Priors
  
  # Hyperpriors describe the community
  # To constrain the model as much as is reasonable, we used constant sd for 
  # all beta psi (occupancy) parameters, and 
  # all beta p (observation) parameters.
  
  # Hyperpriors for the model of occupancy (psi)
  Mu.lpsi ~ dunif(0,1)          # uniform community prior  
  mu.lpsi <- logit(Mu.lpsi)     # logit model link
  sd.lpsi ~ dunif(0, 12)        # bounds of uniform chosen by trial and error
  
  for (lpsii in 2:npsi){
    mu.betalpsi[lpsii] ~ dunif(-5, 5) # mu for each psi param drawn from uniform prior 
  }
  sd.betalpsi  ~ dunif(0, 6) # constant sd for all species on all psi parameters
  
  # prior for elevation polynomial
  # gamma version
  #  mu.betalpsi[1] ~ dunif(0, 6) # the shape parameter gamma- uniform prior
  #  sd.betalpsi1 ~ dunif(0, 6) # the rate parameter gamma- uniform prior
  # lognorm version 
  #mu.betalpsi[1] ~ dunif(-5, 5) # the mu parameter lognorm- uniform prior
  #sd.betalpsi1 ~ dunif(0, 6) # the sd parameter lognorm- uniform prior
  # trunc norm version  
  #  mu.betalpsi[1] ~ dunif(0, 5) # the halfnorm mean -  uniform prior (use same sd as others)
  #  sd.betalpsi1  ~ dunif(0, 6) 
  # no quadratic version
    mu.betalpsi[1] <- 0 # the mu parameter 
    sd.betalpsi1 <- 0 # the sd parameter 
  
  # Hyperpriors for the model of detection (p)
  Mu.lp ~ dunif(0,1)          # uniform community prior
  mu.lp <- logit(Mu.lp)       # logit model link
  sd.lp ~ dunif(0, 5)         # bounds of uniform chosen by trial and error
  
  for (lpi in 1:np){
    mu.betalp[lpi] ~ dunif(-5, 5) # mu for each p param drawn from uniform prior 
  }
  sd.betalp ~ dunif(0, 6) # constant sd for all p parameters
  
  # Priors for species-specific effects in occupancy and detection
  # are derived from these community level distributions
  
  for(k in 1:M){   
    # occupancy 
    lpsi[k] ~ dnorm(mu.lpsi, sd=sd.lpsi)    
    for (lpsii in 2:npsi){
      betalpsi[k, lpsii] ~ dnorm(mu.betalpsi[lpsii], sd=sd.betalpsi)
    }
    # gamma distribution for the polynomial
    #betalpsi[k, 1] ~ dgamma(shape=mu.betalpsi[1], rate=sd.betalpsi1) 
    # lognorm distribution for the polynomial - 
    # betalpsi[k, 1] ~ dlnorm(meanlog=mu.betalpsi[1], sdlog=sd.betalpsi1)
    # truncnorm distribution for the polynomial
    #betalpsi[k, 1] ~ T(dnorm(mean = mu.betalpsi[1], sd = sd.betalpsi1), 0, Inf)
    # no quadratic version
    betalpsi[k, 1] <- 0
    
    lp[k] ~ dnorm(mu.lp, sd=sd.lp)
    for(lpi in 1:np){
      betalp[k, lpi] ~ dnorm(mu.betalp[lpi], sd=sd.betalp)
    }
    
  }       
  
  # the multi-species occupancy model
  
  for(k in 1:M){
    for (i in 1:nSite){
      
      logit(psi[i,k]) <- lpsi[k] + 
        #betalpsi[k,1] * Xoc[i,1] +  # elevation2
        betalpsi[k,2] * Xoc[i,2] +  # elevation1
        betalpsi[k,3] * Xoc[i,3] +  # farm_85
        betalpsi[k,4] * Xoc[i,4] +  # fl_dis
        betalpsi[k,5] * Xoc[i,5] +  # sidi1ha
        betalpsi[k,6] * Xoc[i,6]    # slope
      
      # mu.psi[i,k] <- psi[i,k]       
      # z[i,k] ~ dbern(mu.psi[i,k])       ## occupancy drawn from psi - replaced by dOcc
      
      for(j in 1:nVisits){
        
        logit(p[i,j,k]) <- lp[k] + 
          betalp[k,1] * Xob[i,j,1] + # start2
          betalp[k,2] * Xob[i,j,2] + # date
          betalp[k,3] * Xob[i,j,3] + # start1
          betalp[k,4] * Xob[i,j,4]   # visibility
        
        # mu.p[i,j,k] <- z[i,k] * p[i,j,k]   
        # Y[i,j,k] ~ dbern(mu.p[i,j,k])     ## observations from occupancy and detection (p) - replaced by dOcc
        
      }
      # Marginalized likelihood (dOcc_v from nimbleEcol)
      Y[i,1:nVisits,k] ~ dOcc_v(
        probOcc = psi[i,k], 
        probDet = p[i,1:nVisits,k], 
        len = nVisits)
    }
  }
})
```

# Specify model: code, data (observed), constants (set), and inits (starting value for estimated values)

```{r import data}
nimData <- readRDS(data_input)

y_all <- nimData$y_all
farmSites <- nimData$farmSites
OccVarsFarm <- nimData$OccVarsFarm
ObsVarsFarm <- nimData$ObsVarsFarm
Nadd <- 11 # observed 116 of an estimated 127 species = add 11

y <- y_all[(dimnames(y_all)[[1]] %in% farmSites), , ]  # [site, visit, species]

# remove species for which no observations were made in this habitat
y <- y[,, apply(y, c(3), sum) > 0]

nSite <- dim(y)[1]
nVisits <- dim(y)[2]
Nobs <- dim(y)[3]
M <- Nobs + Nadd
Y <- abind::abind(y, array(0, dim = c(dim(y)[1:2], Nadd)))  ## add the 'missing' species - not seen so = 0, 'absence'

# occupancy variables [site, vars]
# 1"elevation2" 2"elevation1" 3"farm_85" 4"fl_dis" 5 "sidi1ha" 6"slope
Xoc <- OccVarsFarm %>% select(-point_id) %>% as.matrix() %>% array(dim = dim(.)) 
names(OccVarsFarm)
Xoc <- Xoc[,c(2,1,3:6)]  

# observation variables [site, visit, vars]
# 1"start2", 2"date",3"start1",4"visibility"
Xob <- array(NA, dim = c(nSite, nVisits, dim(ObsVarsFarm)[2]-2)) 
Xob[,,1] <- ObsVarsFarm %>% arrange(round, point_id) %>% pull(date)
Xob[,,2] <- ObsVarsFarm %>% arrange(round, point_id) %>% pull(start1)
Xob[,,3] <- ObsVarsFarm %>% arrange(round, point_id) %>% pull(start2)
Xob[,,4] <- ObsVarsFarm %>% arrange(round, point_id) %>% pull(visibility)
Xob <- Xob[,,c(3,1,2,4)] 

npsi <- dim(Xoc)[2]
np <- dim(Xob)[3]

# Initial values - match up with the number of observed vars used in the model
lpsi <- rnorm(n = M)
betalpsi <- array(rnorm(n = M * npsi), dim = c(M, npsi))
# betalpsi[1:M, 1] <- rgamma(n = M, shape = 1, rate = 1) # correct betalpsi to be rgamma
# betalpsi[1:M, 1] <- rlnorm(n = M, meanlog=0, sdlog=1) # lognorm version
# betalpsi[1:M, 1] <- abs(betalpsi[1:M, 1]) # truncnorm version
 betalpsi[1:M, 1] <- 0 # no quadratic version

lp <- rnorm(n = M)
betalp <- array(rnorm(n = M * np), dim = c(M, np))

# list required variables
data <- list(Y = Y, 
             Xoc = Xoc,
             Xob = Xob)  
constants <- list(nSite = nSite, 
                  nVisits = nVisits, 
                  Nobs = Nobs,
                  Nadd = Nadd, 
                  M = M,
                  npsi = npsi,
                  np = np)
inits <- list(
  lpsi = lpsi, 
  betalpsi = betalpsi, 
  lp = lp, 
  betalp = betalp 
)
params1 <- c("lpsi", "betalpsi", "lp", "betalp", "psi", "p")#,
             # "mu.betalpsi", "sd.betalpsi1")  # gamma
             #  "mu.betalpsi", "sd.betalpsi1") # lnorm
             #"mu.betalpsi", "sd.betalpsi1")  # truncnorm
```

## run MCMC
```{r run nimbleMCMC}

# MCMC settings - development -> final numbers of parameters
# ni <- 10      ;   nt <- 1    ;   nb <- 0       ;   nc <- 3
# ni <- 2500   ;   nt <- 5   ;   nb <- 0   ;   nc <- 3
ni <- 30000   ;   nt <- 5   ;   nb <- 10000   ;   nc <- 3
# (ni - nb)/nt # samples per chain = 4000 in the final models

samples <- nimbleMCMC(
  code = occupancy_code,
  data = data,
  constants = constants,
  inits = inits,
  monitors = params1,
  thin = nt,
  niter = ni,
  nburnin = nb,
  nchains = nc)
```

# model diagnostics
```{r MCMCvis}
library(MCMCvis)
pg1 <- c(#"mu.betalpsi", "sd.betalpsi1", 
         "lpsi", "betalpsi", "lp", "betalp") 

# summary of estimates --- we want Rhat < 1.1, and n.eff >100
mysum <- MCMCsummary(samples, 
                     params = pg1, 
                     round = 2)

mysum[which(mysum$Rhat>1.1),] # %>% dimnames()
mysum[which(mysum$n.eff<100),]# %>% dimnames() 

# trace plots --- show mixing/convergence
MCMCtrace(samples, 
          params = pg1,     # check all
          #params = "lpsi", # check one
          ind = TRUE, 
          Rhat = TRUE, n.eff = TRUE,
          filename = paste0("Results/", version_folder, 
                            "plots_traceDensity_pg1_farm_", 
                            "_trial", trialName,
                            format(Sys.time(), "%Y%m%d"), ".pdf"))

# caterpillar plots - choose parameter option
MCMCplot(samples, 
         params = 'betalpsi\\[\\d{1,3}[,][ ][1]\\]', ISB = FALSE, exact = FALSE,  
         #params = 'betalp\\[\\d{1,3}[,][ ][1]\\]', ISB = FALSE, exact = FALSE,  
         #params = "lp",
         #params = "lpsi",
         ci = c(50, 90), 
         HPD = FALSE)  # plots equal tailed credible intervals
```


```{r save results and data inputs}
saveRDS(samples, paste0(results_folder, version_folder, 
                        "BirdOccMod_dOcc_samples_farm_", 
                        format(Sys.time(), "%Y%m%d"), 
                        "_trial", trialName, ".RDS"))
saveRDS(list(data,constants), paste0("Data/Farm_nimbleOccData_",data_version,"_", 
                                     format(Sys.time(), "%Y%m%d"), ".RDS"))
```
